Complete project....Rebase went wrong.
I did not know how to rebase. I still mostly don't know how to rebase.
Thankfully, VS Code saved my ass by having a timeline of my files
independent of git...Just imagine there being like 30 more commits
between this one and the last one.

The original images I committed had metadata I did not want public,
so I thought I would try to remove the commit since it was just a
couple commits ago and rebasing seemed easier than filter-branch.

I deleted the commit, but wasn't totally clear on the conflicts. But
everything seemed correct in the log, so I just continued working on the
project...inside of the interactive rebase's detached head.

I finish the project after a couple hours and reviewed it to make sure
everything looks good. I then try to push to remote and it says it can't
push when a rebase is in progress. I see that I'm not in the main branch
and try to jump to it to see what it looks like, losing all commits in
the detached head in the process, I think.

After restoring the completed files, I try to rebase again, but I'm not
really sure how to deal with conflicts, so I just ended up dropping all
commits after adding the original images and then restoring the
completed files and cleaned up images.

I am very fortunate that this happened on a solo educational project
and not on a work project that involves other developers.